[Finishes #119893127] Create an AuthContainer
This now handles auth actions - login and signup
