+++ client/src/routes/index.js
@@ -8,8 +8,6 @@ import configureStore from '../store/configureStore';
 
 import MainContainer from '../containers/Main/MainContainer';
 import HomeContainer from '../containers/Home/HomeContainer';
-import LoginContainer from '../containers/Auth/LoginContainer';
-import SignupContainer from '../containers/Auth/SignupContainer';
 
 const store = configureStore();
 
@@ -18,8 +16,6 @@ const routes = (
     <Router history={hashHistory}>
       <Route path='/' component={MainContainer}>
         <IndexRoute component={HomeContainer} />
-        <Route path='/login' component={LoginContainer}></Route>
-        <Route path='/signup' component={SignupContainer}></Route>
       </Route>
     </Router>
   </Provider>
